Define WidgetInterface
* VuFind/Controller/Widget/WidgetInterface.php: New interface.

A Widget is a special type of controller plugin comparable to a web
portal Portlet. It is a self-contained application component that is
interacted and interacts with the main application. The key feature of
a Widget is its capability to represent its internal state as a set of
query parameters.