Update module/VuFind/src/VuFind/Controller/Plugin/Holds.php
For the Aleph Driver, it's possible that the item ID (id from route) and the holding ID (id from query) are different. With the += operator for merging, the id attribute from $keyValueArray was not used and an invalid REST url was sent to the ILS.
I'm not sure why the record id is loaded from the route here. With this fix, it's only used as a fallback when not defined in the query as a HMACKeys-variable of the ILS.

Aleph uses id:item_id which causes the conflict here.